+++ CHANGELOG.md
@@ -10,14 +10,22 @@ All notable changes to this project will be documented in this file.
 
 [#647]: https://github.com/stackabletech/operator-rs/pull/647
 
+### Changed
+
+- Bump `kube` to `0.86.0` and Kubernetes version to `1.28` ([#648]).
+
+[#648]: https://github.com/stackabletech/operator-rs/pull/648
+
 ## [0.48.0] - 2023-08-18
 
 ### Added
 
 - Add `PodBuilder::termination_grace_period_seconds` ([#641]).
 - Add support for adding `lifecycle`s to `ContainerBuilder` ([#641]).
+- Add support for tls pkcs12 password to secret operator volume builder ([#645]).
 
 [#641]: https://github.com/stackabletech/operator-rs/pull/641
+[#645]: https://github.com/stackabletech/operator-rs/pull/645
 
 ## [0.47.0] - 2023-08-16